requirements.txt
youtube-transcript-api
openai
python-dotenv
agents
streamlit